+++ webknossos/webknossos/geometry/__init__.py
@@ -1,7 +1,7 @@
 # ruff: noqa: F401 imported but unused
 
 from .bounding_box import BoundingBox
-from .mag import Mag
+from .mag import Mag, MagLike
 from .nd_bounding_box import NDBoundingBox
 from .normalized_bounding_box import NormalizedBoundingBox
 from .vec3_int import Vec3Int, Vec3IntLike