+++ webknossos/webknossos/geometry/normalized_bounding_box.py
@@ -6,6 +6,7 @@ from .nd_bounding_box import NDBoundingBox
 from .vec_int import VecInt
 
 _DEFAULT_AXIS_ORDER = {"c": 0, "x": 1, "y": 2, "z": 3}
+_DEFAULT_AXIS_ORDER_MISSING_C = {"x": 1, "y": 2, "z": 3}
 
 
 @attr.frozen
@@ -137,7 +138,11 @@ class NormalizedBoundingBox(NDBoundingBox):
         if (
             bbox.get("channelIndex", 0) == 0
             and ("additionalAxes" not in bbox or bbox["additionalAxes"] == [])
-            and ("axisOrder" not in bbox or bbox["axisOrder"] == _DEFAULT_AXIS_ORDER)
+            and (
+                "axisOrder" not in bbox
+                or bbox["axisOrder"]
+                in (_DEFAULT_AXIS_ORDER, _DEFAULT_AXIS_ORDER_MISSING_C)
+            )
         ):
             # Delegate to BoundingBox.from_wkw_dict, if only 3d
             from .bounding_box import BoundingBox
@@ -180,7 +185,8 @@ class NormalizedBoundingBox(NDBoundingBox):
         axes = [
             axis
             for axis in axes
-            if axis.name in bbox.get("axisOrder", {}).keys()  # in axisOrder
+            if axis.name
+            in bbox.get("axisOrder", _DEFAULT_AXIS_ORDER).keys()  # in axisOrder
             or any(
                 a["name"] == axis.name for a in bbox.get("additionalAxes", [])
             )  # or in additionalAxes
@@ -207,13 +213,16 @@ class NormalizedBoundingBox(NDBoundingBox):
             dict: A json dictionary representing the bounding box.
         """
         if self.axes == ("c", "x", "y", "z"):
-            return {
+            out = {
                 "topLeft": self.topleft_xyz.to_list(),
                 "width": self.size.x,
                 "height": self.size.y,
                 "depth": self.size.z,
                 "numChannels": self.size.c,
             }
+            if self.topleft.c != 0:
+                out["channelIndex"] = self.topleft.c
+            return out
 
         topleft = [None, None, None]
         width, height, depth = None, None, None
@@ -257,8 +266,8 @@ class NormalizedBoundingBox(NDBoundingBox):
             minimal_axis_order = {
                 axis: index
                 for axis, index in axis_order.items()
-                if axis not in _DEFAULT_AXIS_ORDER
-                or (ndim - len(_DEFAULT_AXIS_ORDER) - _DEFAULT_AXIS_ORDER[axis])
+                if axis in _DEFAULT_AXIS_ORDER
+                and (ndim - len(_DEFAULT_AXIS_ORDER) - _DEFAULT_AXIS_ORDER[axis])
                 != index
             }
             return minimal_axis_order